Delete Chofer

+++ src/Controller/ChoferController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Chofer;
+use App\Entity\Combi;
 use App\Form\ChoferType;
 use App\Repository\ChoferRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -83,10 +84,21 @@ class ChoferController extends AbstractController
      */
     public function delete(Request $request, Chofer $chofer): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$chofer->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($chofer);
-            $entityManager->flush();
+        $repository = $this->getDoctrine()->getRepository(Combi::class);
+        $combi = $repository->findOneBy(['chofer' =>  $chofer ]);
+
+        if ( !($combi ) ){
+        
+            if ($this->isCsrfTokenValid('delete'.$chofer->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($chofer);
+                $entityManager->flush();
+            }
+
+        } else {
+
+            $this->addFlash('failed', 'Ese chofer se encuentra en uso!');
+
         }
 
         return $this->redirectToRoute('chofer_index');
